Move projection out of types module
