+++ crates/control/Cargo.toml
@@ -16,6 +16,7 @@ kinematics = { workspace = true }
 log = { workspace = true }
 nalgebra = { workspace = true }
 ordered-float = { workspace = true }
+projection = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 serialize_hierarchy = { workspace = true }
