+++ crates/projection/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "projection"
+version = "0.1.0"
+edition = "2021"
+license = "GPL-3.0-only"
+homepage = "https://github.com/hulks/hulk"
+
+[dependencies]
+nalgebra = { workspace = true }
+thiserror = { workspace = true }
+types = { workspace = true }
+
+[dev-dependencies]
+approx = { workspace = true }
