+++ crates/types/src/lib.rs
@@ -22,6 +22,7 @@ mod game_controller_state;
 mod geometry;
 pub mod grayscale_image;
 pub mod hardware;
+pub mod horizon;
 mod image_segments;
 mod initial_pose;
 mod joints;
@@ -70,7 +71,7 @@ pub mod ycbcr422_image;
 pub use ball::{Ball, CandidateEvaluation};
 pub use ball_position::BallPosition;
 pub use buttons::Buttons;
-pub use camera_matrix::{CameraMatrices, CameraMatrix, Horizon, ProjectedFieldLines};
+pub use camera_matrix::{CameraMatrices, CameraMatrix, ProjectedFieldLines};
 pub use camera_position::CameraPosition;
 pub use color::{Intensity, Rgb, RgbChannel, YCbCr422, YCbCr444};
 pub use cycle_time::CycleTime;
